2018.12.27 15:00 表结构更新

+++ smartGym-backEnd/smartGym-backEnd-dao/src/main/java/cn/smartGym/mapper/SmartgymApplicationsMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface SmartgymApplicationsMapper {
-    long countByExample(SmartgymApplicationsExample example);
+    int countByExample(SmartgymApplicationsExample example);
 
     int deleteByExample(SmartgymApplicationsExample example);
